Connect leaderboard table to world level zone dropdowns

+++ Assets/Scripts/LeaderBoard/UserScore.cs
@@ -4,8 +4,8 @@ using System;
 public class UserScore
 {
     public String username;
-    public float score;
-    public UserScore(String username, float score)
+    public int score;
+    public UserScore(String username, int score)
     {
         this.username = username;
         this.score = score;
